requirements.txt
Pillow>=10.0.0
gradio>=4.44.0
